+++ tests/test_monitor.py
@@ -298,13 +298,22 @@ class TestMonitor(object):
 
 class TestMonitorObserver(object):
 
-    def receive_event(self, action, device):
+    def callback(self, device):
+        self.events.append(device)
+        if len(self.events) >= 2:
+            self.observer.send_stop()
+
+    def event_handler(self, action, device):
         self.events.append((action, device))
         if len(self.events) >= 2:
             self.observer.send_stop()
 
-    def make_observer(self, monitor):
-        self.observer = MonitorObserver(monitor, self.receive_event)
+    def make_observer(self, monitor, use_deprecated=False):
+        if use_deprecated:
+            self.observer = pytest.deprecated_call(
+                MonitorObserver, monitor, event_handler=self.event_handler)
+        else:
+            self.observer = MonitorObserver(monitor, callback=self.callback)
         return self.observer
 
     def setup(self):
@@ -313,19 +322,30 @@ class TestMonitorObserver(object):
     def teardown(self):
         self.events = None
 
+    def test_deprecated_handler(self, fake_monitor, fake_monitor_device):
+        observer = self.make_observer(fake_monitor, use_deprecated=True)
+        observer.start()
+        fake_monitor.trigger_event()
+        fake_monitor.trigger_event()
+        # wait a second for the tests to finish, and kill the observer if
+        # it is still alive then
+        observer.join(1)
+        if observer.is_alive():
+            observer.stop()
+        assert self.events == [(None, fake_monitor_device)] * 2
+
     def test_fake(self, fake_monitor, fake_monitor_device):
         observer = self.make_observer(fake_monitor)
         observer.start()
         fake_monitor.trigger_event()
         fake_monitor.trigger_event()
-        # fake one second for the tests to finish
+        # wait a second for the tests to finish
         observer.join(1)
         # forcibly quit the thread if it is still alive
         if observer.is_alive():
             observer.stop()
         # check that we got two events
-        assert self.events == [(None, fake_monitor_device),
-                               (None, fake_monitor_device)]
+        assert self.events == [fake_monitor_device] * 2
 
     @pytest.mark.privileged
     def test_real(self, context, monitor):
@@ -339,7 +359,6 @@ class TestMonitorObserver(object):
         observer.join(2)
         if observer.is_alive():
             observer.stop()
-        assert [e[0] for e in self.events] == ['add', 'remove']
-        assert [e[1].action for e in self.events] == ['add', 'remove']
-        for _, device in self.events:
+        assert [d.action for d in self.events] == ['add', 'remove']
+        for device in self.events:
             assert device.device_path == '/devices/virtual/net/dummy0'